+++ package.json
@@ -45,6 +45,7 @@
     "react-katex": "^3.0.1",
     "react-number-format": "^5.4.2",
     "react-redux": "^9.1.2",
+    "react-router": "^7.0.2",
     "react-toastify": "^10.0.6",
     "tailwind-merge": "^2.4.0",
     "tailwindcss-animate": "^1.0.7",
